Put the input message on fixed-width fields and bound its key count
The inputState payload wrote the key count as a size_t and each key code as a bare int. size_t is eight
bytes on a 64-bit host and four on a 32-bit one, so a client and server of different widths desynchronize
on the first input message and stay wrong for every field after it. The replication path already writes
uint32_t explicitly; the input path now matches, with the count as uint32 and each key as int32.

The server also sized its key vector straight from the count it had just read, so a client sending a
count of a billion asked for a multi-gigabyte allocation before anything validated it. The drain loop
catching std::exception makes that survivable rather than fatal, but the allocation is still attempted,
and a value just under the failure threshold is a cheap way to make the server thrash. The count is now
checked against what is left of the payload first - a message cannot hold more key codes than it has
bytes for.

The bound stays deliberately loose (it counts the trailing mouse block as if it were key bytes) so that a
client predating mouse input still degrades to "no mouse" rather than being rejected.

+++ source/apps/editor/EditorApp.cpp
@@ -482,10 +482,10 @@ void EditorApp::sendInput()
 
   net::Message message(net::MessageType::inputState);
   message.write(snapshot.focused);
-  message.write(snapshot.keys.size());
+  message.write(static_cast<uint32_t>(snapshot.keys.size()));
   for (const auto& key : snapshot.keys)
   {
-    message.write(key);
+    message.write(static_cast<int32_t>(key));
   }
 
   message.write(snapshot.mouseX);
